@@ -25,12 +25,13 @@
   "devDependencies": {
     "dotenv-cli": "^6.0.0",
     "eslint-config-custom": "*",
+    "nodemon": "^2.0.20",
     "prettier": "latest",
     "turbo": "latest",
     "typescript": "^4.9.4"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=18.0.0"
   },
   "packageManager": "npm@8.19.2",
   "dependencies": {
