@@ -1,4 +1,3 @@
-"use strict";
-exports.__esModule = true;
-var client_1 = require("@prisma/client");
-exports["default"] = client_1.PrismaClient;
+import { PrismaClient } from "@prisma/client";
+
+export default PrismaClient;
