@@ -2,12 +2,12 @@ const tailwindcss = require("tailwindcss");
 const autoprefixer = require("autoprefixer");
 
 const config = {
-  plugins: [
-    //Some plugins, like tailwindcss/nesting, need to run before Tailwind,
-    tailwindcss(),
-    //But others, like autoprefixer, need to run after,
-    autoprefixer,
-  ],
+    plugins : [
+    // Some plugins, like tailwindcss/nesting, need to run before Tailwind,
+        tailwindcss(),
+        // But others, like autoprefixer, need to run after,
+        autoprefixer,
+    ],
 };
 
 module.exports = config;
