@@ -1,8 +1,6 @@
 <script lang="ts">
-    import JSONTree from "svelte-json-tree";
-    
     export let transaction:Record<string, any> = {};
 </script>
 <div class="card">
-    <JSONTree value={transaction} />
+    <pre>{JSON.stringify(transaction)}</pre>
 </div>
