@@ -1,7 +1,6 @@
-<!-- The layout for /wallet /tx /token -->
+ <!-- The layout for /wallet /tx /token -->
 <div class="min-h-screen px-3 pt-32">
     <div class="w-full max-w-xl mx-auto">
         <slot></slot>
     </div>
 </div>
-
