@@ -1,1 +1,11 @@
-<!-- empty page required else 404 -->
+<script>
+
+    import { Server } from "@sveltejs/kit";
+
+</script>
+<!--
+    Empty page required else this route is a 404
+
+    The logic of this route is in layout.server.ts
+    which determines whether to redirect to /wallet, /tx, or /token
+-->
